refactor: move post comments endpoint
